% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyExamples.R
\name{ARCoeffMap}
\alias{ARCoeffMap}
\title{Generate the AR coefficient map.}
\usage{
ARCoeffMap(Ns = 64)
}
\arguments{
\item{Ns}{Number of pixels of the result in one direction. The resulting
picture will have Ns x Ns pixels.}
}
\value{
A list containing x and y, the coordinates of the grid and
       z, a matrix of dimensions Ns x Ns giving the AR coefficients map.
}
\description{
Generate the AR coefficient map.
}

