% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawContour.R
\name{DrawContour}
\alias{DrawContour}
\title{Draws the contour {f=c}.}
\usage{
DrawContour(..., level, col, lty = 1)
}
\arguments{
\item{...}{An image. Either as a list with components x,y and z or as vectors
x and y and a matrix z of dimensions c(length(x),length(y)).}

\item{level}{The level of the contour.}

\item{col}{Color of the contour.}

\item{lty}{Line type for the contour.}
}
\description{
Draws the contour {f=c}.
}

