% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageMap.R
\name{ImageMap}
\alias{ImageMap}
\title{Plots an image on a map of the world.}
\usage{
ImageMap(lon, lat, img, mask = NULL, xlab = "Longitude",
  ylab = "Latitude", ...)
}
\arguments{
\item{lon}{Longitude.}

\item{lat}{Latitude.}

\item{img}{The image to plot as a matrix of dimensions 
c(length(lon),length(lat)).}

\item{mask}{Matrix of dimensions c(length(lon),length(lat)) defining a mask 
to cut out of the picture.}

\item{xlab}{Label for the x-axis passed to fields::image.plot. Default is "Longitude".}

\item{ylab}{Label for the y-axis passed to fields::image.plot. Default is "Latitude".}

\item{...}{Additional graphical parameters passed to fields::image.plot.}
}
\description{
Plots an image on a map of the world.
}

