% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCope.R
\name{PlotCope}
\alias{PlotCope}
\title{Plots CoPE sets.}
\usage{
PlotCope(cope, plot.taylor = FALSE, use.true.function = FALSE,
  map = FALSE, ...)
}
\arguments{
\item{cope}{An object of class cope to be plotted.}

\item{plot.taylor}{Boolean indicating whether the CoPE sets with the threshold 
obtained by Taylor's method should be plotted. Default is
FALSE.}

\item{use.true.function}{Boolean indicating whether the threshold obtained 
from the true function should be used. Default is 
FALSE.}

\item{map}{If TRUE plot the cope set on a map of the world. The coordinates 
in this case are interpreted as longitude and latitude.}

\item{...}{Additional graphical parameters passed to fields::image.plot.}
}
\description{
Plots CoPE sets.
}

