% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetContour.R
\name{SubsetContour}
\alias{SubsetContour}
\title{Determine whether {f1>=c1} a subset of {f2>=c2}.}
\usage{
SubsetContour(x, y, f1, c1, f2, c2)
}
\arguments{
\item{x}{x-Coordinates of the grid on which the functions are observed.}

\item{y}{y-Coordinates of the grid on which the functions are observed.}

\item{f1}{A matrix of dimension c(length(x),length(y)) defining the first 
function.}

\item{c1}{The first level.}

\item{f2}{A matrix of dimension c(length(x),length(y)) defining the second 
function.}

\item{c2}{The second level.}
}
\value{
A boolean indicating whether {f1>=c1} a subset of {f2>=c2}.
}
\description{
Determine whether {f1>=c1} a subset of {f2>=c2}.
}

