% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ToyNoise1}
\alias{ToyNoise1}
\title{Generate a realization of the Toy Noise 1.}
\usage{
ToyNoise1(n = 1)
}
\arguments{
\item{n}{The number of realizations to produce.}
}
\value{
A list containing x and y, the coordinates of the grid and
       z and array of dimensions c(64,64,n) giving n reallizations of the
       Toy Noise 1.
}
\description{
Generate a realization of the Toy Noise 1.
}

