% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyExamples.R
\name{ToyNoise1Presmooth}
\alias{ToyNoise1Presmooth}
\title{Generate a realization of the Toy Noise 1 before smoothing.}
\usage{
ToyNoise1Presmooth(n = 1, Ns = 64, model = list(), theta = 0.1,
  l1 = 1, l2 = 4, tau = 12)
}
\arguments{
\item{n}{The number of realizations to produce.}
}
\value{
A list containing x and y, the coordinates of the grid and
       z and array of dimensions c(64,64,n) giving n reallizations of the 
       Toy Noise 1 before smoothing.
}
\description{
Generate a realization of the Toy Noise 1 before smoothing.
}

