% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyExamples.R
\name{ToyNoiseMap}
\alias{ToyNoiseMap}
\title{Generate an AR sequence with the ToyFUN as base noise and the AR coefficients
given by the ARCoeffMap.}
\usage{
ToyNoiseMap(n = 1, ToyFUN, ...)
}
\arguments{
\item{n}{Length of sequence.}

\item{ToyFUN}{Base noise to build the sequence from.}

\item{...}{Additional parameters passed to ToyFUN.}
}
\value{
A list containing x and y, the coordinates of the grid and
       z and array of dimensions c(64,64,n) giving n realizations of the 
       Toy Noise.
}
\description{
Generate an AR sequence with the ToyFUN as base noise and the AR coefficients
given by the ARCoeffMap.
}

