% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyExamples.R
\name{ToySlope}
\alias{ToySlope}
\title{The toy slope.}
\usage{
ToySlope(ImRange = c(0, 1), NPixel = 64)
}
\arguments{
\item{ImRange}{A vector with two components giving the range of the region on
which the Toy Slope is to be computed.}

\item{NPixel}{Number of pixels of the result in one direction. The resulting
picture will have NPixel x NPixel pixels.}
}
\value{
A list with components "x", "y" and "z". Here, x and y are the 
        coordinates of the grid and z is matrix of dimensions 
        c(NPixel,NPixel) giving the Toy Signal.
}
\description{
The toy slope.
}

